feat: add triangle links

+++ Runtime/Core/Mesh/Generators/IcosahedronGen.cs
@@ -17,20 +17,20 @@ namespace UnityEngine.GsplEdit
         {
             public Vector3 center;
             public fixed float vertices[12 * 3];
-            public fixed uint indices[60];
+            public fixed int indices[60];
             public float opacity;
             public Vector3 boundMin;
             public Vector3 boundMax;
 
             public static int GetSize() {
-                return sizeof(float) * (3 + 12*3 + 1 + 3 + 3) + sizeof(uint) * 60;
+                return sizeof(float) * (3 + 12*3 + 1 + 3 + 3) + sizeof(int) * 60;
             }
         }
 
         public Settings m_Settings = new();
         public ComputeShader m_IcosahedronComputeShader;
 
-        public unsafe override void Generate(SharedComputeContext context, ref Vertex[] vertexList, ref uint[] indexList) {
+        public unsafe override void Generate(SharedComputeContext context, ref Vertex[] vertexList, ref int[] indexList) {
 
             int splatCount = context.splatData.splatCount;
             int itemsPerDispatch = 65535;
@@ -84,7 +84,7 @@ namespace UnityEngine.GsplEdit
                 // Add indices
                 for (int j = 0; j < 60; j++)
                 {
-                    indexList[indexIndex] = (uint)icosahedron.indices[j];
+                    indexList[indexIndex] = icosahedron.indices[j];
                     indexIndex++;
                 }
             }
